Revert "Feature to directly covert image from URL."
